requirements.txt
python-telegram-bot==21.5
requests
beautifulsoup4